Docker/python-worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
PyNaCl==1.5.0
